# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BusinessShield API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /websites:
    get:
      operationId: listWebsites
      tags: [websites]
      summary: List all websites for the authenticated user
      responses:
        "200":
          description: List of websites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Website"
    post:
      operationId: createWebsite
      tags: [websites]
      summary: Add a new website to monitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebsiteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"
  /websites/{id}:
    get:
      operationId: getWebsite
      tags: [websites]
      summary: Get a website by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Website
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"
        "404":
          description: Not found
    patch:
      operationId: updateWebsite
      tags: [websites]
      summary: Update a website
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebsiteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"
    delete:
      operationId: deleteWebsite
      tags: [websites]
      summary: Delete a website
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Scans ───────────────────────────────────────────────────────────────────
  /websites/{id}/scans:
    get:
      operationId: listScans
      tags: [scans]
      summary: List scans for a website
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of scans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scan"
    post:
      operationId: triggerScan
      tags: [scans]
      summary: Trigger a new scan for a website
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanTriggerInput"
      responses:
        "201":
          description: Scan queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
  /scans/{id}:
    get:
      operationId: getScan
      tags: [scans]
      summary: Get a scan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"

  # ─── Findings ────────────────────────────────────────────────────────────────
  /scans/{id}/findings:
    get:
      operationId: listFindings
      tags: [findings]
      summary: List findings for a scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of findings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanFinding"

  # ─── Monitoring ──────────────────────────────────────────────────────────────
...
  /team/members:
    get:
      operationId: listTeamMembers
      tags: [team]
      summary: List team members
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: inviteTeamMember
      tags: [team]
      summary: Invite a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInviteInput"
      responses:
        "201":
          description: Invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team/members/{id}:
    delete:
      operationId: removeTeamMember
      tags: [team]
      summary: Remove a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ─── Developer / API Keys ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Website:
      type: object
      required: [id, userId, url, name, status, riskScore, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        url:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, paused, error]
        riskScore:
          type: ["integer", "null"]
        complianceScore:
          type: ["integer", "null"]
        lastScannedAt:
          type: ["string", "null"]
        nextScanAt:
          type: ["string", "null"]
        monitoringEnabled:
          type: boolean
        scanFrequency:
          type: ["string", "null"]
          enum: [daily, weekly, monthly, "null"]
        sslExpiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    WebsiteInput:
      type: object
      required: [url, name]
      properties:
        url:
          type: string
          format: uri
        name:
          type: string
        monitoringEnabled:
          type: boolean
        scanFrequency:
          type: string
          enum: [daily, weekly, monthly]
    WebsiteUpdate:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        status:
          type: string
          enum: [active, paused]
        monitoringEnabled:
          type: boolean
        scanFrequency:
          type: string
          enum: [daily, weekly, monthly]

    # ── Scans ─────────────────────────────────────────────────────────────────
    Scan:
      type: object
      required: [id, websiteId, status, scanTypes, createdAt]
      properties:
        id:
          type: integer
        websiteId:
          type: integer
        status:
          type: string
          enum: [queued, running, completed, failed]
        scanTypes:
          type: array
          items:
            type: string
        riskScore:
          type: ["integer", "null"]
        complianceScore:
          type: ["integer", "null"]
        summary:
          type: ["string", "null"]
        findingsCount:
          type: ["integer", "null"]
        criticalCount:
          type: ["integer", "null"]
        highCount:
          type: ["integer", "null"]
        mediumCount:
          type: ["integer", "null"]
        lowCount:
          type: ["integer", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ScanTriggerInput:
      type: object
      properties:
        scanTypes:
          type: array
          items:
            type: string
          default: [accessibility, ssl, security, privacy, seo, performance]

    # ── Findings ──────────────────────────────────────────────────────────────
    ScanFinding:
      type: object
      required: [id, scanId, category, severity, title, description, createdAt]
      properties:
        id:
          type: integer
        scanId:
          type: integer
        category:
          type: string
          enum: [accessibility, ssl, security, privacy, seo, performance, legal]
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        title:
          type: string
        description:
          type: string
        recommendation:
          type: ["string", "null"]
        codeSnippet:
          type: ["string", "null"]
        wcagCriteria:
          type: ["string", "null"]
        element:
          type: ["string", "null"]
        isFixed:
          type: boolean
        createdAt:
          type: string

    # ── Monitoring ────────────────────────────────────────────────────────────
...
    TeamMember:
      type: object
      required: [id, userId, email, role, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, admin, member, viewer]
        status:
          type: string
          enum: [active, invited, suspended]
        createdAt:
          type: string
    TeamInviteInput:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, member, viewer]

    # ── Developer / API Keys ──────────────────────────────────────────────────────
...